Font/Hero/Footer
I added in some javascript to make a slider for the hero page like the official 8bitdo site, and also decided to use the same images and try to be as close as i can to official site without taking their code. I also fixed up the footer at the bottom and changed the font. I still need to make sure everything is responsive, probably redo the hero slider with text and buttons and refine navbar with responsiveness and a hamburger menu